Move cron into other and continue to add godoc
