+++ payload/download/download.go
@@ -1,3 +1,7 @@
+// Download and execute payloads.
+//
+// The download package contains all the code for download and execute payloads. Unlike the other payloads
+// this package is necessarily OS dependent for both the download and execution portions.
 package download
 
 type Download interface{}
