+++ payload/other/other_test.go
@@ -1,11 +1,11 @@
-package payload
+package other
 
 import (
 	"testing"
 )
 
 func TestEncodeSelfRemovingCron(t *testing.T) {
-	cron, xploit := SelfRemovingCron("root", "/etc/cron.d/hi", "/tmp/test", "id")
+	cron, xploit := Cron.SelfRemovingCron("root", "/etc/cron.d/hi", "/tmp/test", "id")
 
 	if cron != "* * * * * root /bin/sh /tmp/test\n" {
 		t.Fatal(cron)
